Move summarizer-related code from containerRuntime.ts
