+++ packages/runtime/container-runtime/src/summary/summarizerHeuristics.ts
@@ -6,8 +6,7 @@
 import { Timer } from "@fluidframework/core-utils/internal";
 import { ITelemetryLoggerExt } from "@fluidframework/telemetry-utils/internal";
 
-import { ISummaryConfigurationHeuristics } from "../containerRuntime.js";
-
+import type { ISummaryConfigurationHeuristics } from "./summarizer.js";
 import {
 	ISummarizeAttempt,
 	ISummarizeHeuristicData,
